requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.7; python_version >= '3.6'
bleach==5.0.0
blinker==1.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dnspython==2.2.1; python_version >= '3.6' and python_version < '4.0'
dominate==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.2.1
flask-admin==1.6.0
flask-bootstrap==3.3.7.1
flask-login==0.5.0
flask-mail==0.9.1
flask-migrate==2.7.0
flask-script==2.0.5
flask-sqlalchemy==2.5.1
flask-uploads==0.2.1
flask-wtf==1.0.1
flask==1.1.4
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
idna==3.3; python_version >= '3.5'
importlib-metadata==4.11.3; python_version < '3.10'
importlib-resources==5.7.1; python_version < '3.9'
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mako==1.2.0; python_version >= '3.7'
markupsafe==2.0.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
python-dotenv==0.20.0
setuptools==62.2.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.36; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
visitor==0.1.3
webencodings==0.5.1
werkzeug==0.16.0
wtforms[email]==3.0.1
zipp==3.8.0; python_version >= '3.7'
